requirements.txt
PyYAML~=6.0.3
GitPython~=3.1.45
groq~=0.32.0
python-dotenv~=1.1.1